requirements.txt
google.adk
protobuf
python-dotenv
pyfiglet
termcolor
atproto
finnhub-python
pandas
ta
yfinance
fastmcp